feat: unify generateTspec with NestJS mode and extract common schema builder
- Add generator/schemaBuilder.ts with common schema building logic
- Update nestjs/openapiGenerator.ts to use shared schemaBuilder
- Integrate NestJS mode into generateTspec (nestjs: true option)
- Deprecate generateNestTspec in favor of generateTspec with nestjs flag
- Deprecate GenerateNestTspecOptions type alias

BREAKING CHANGES:
- generateNestTspec is now deprecated, use generateTspec({ nestjs: true }) instead
- GenerateNestTspecOptions is now deprecated, use Tspec.GenerateParams instead
